Add sql pool configuration parameters to nova.conf
Set the sql pool configuration explicitly in nova.conf.
Min pool size is 1 (default), max pool size is 10 (default=5)
The overflow value is set to 50. An overflow connection is created
when all connection in the pool are in use. To avoid a connection storm
we limit the overflow value to 50 instead of setting it to unlimited.

Fixes SMBCAP-4852

Signed-off-by: Christoph Kassen <c.kassen@telekom.de>